refactor(config): rename permissions to oAuthScopes, move shared fn into config package
