+++ packages/application-config/src/build-oidc-scope.ts
@@ -1,28 +1,24 @@
-// TODO: this is copied from the `application-shell` package.
-// Find a way to expose those, or extract them into a shared place.
+import type { ApplicationOidcForDevelopmentConfig } from '@commercetools-frontend/constants';
 
 import { OIDC_CLAIMS } from './constants';
 
 type BuilOidcScopeOptions = {
-  projectKey?: string | null;
-  permissions?: {
-    view: string[];
-    manage: string[];
-  };
-  teamId?: string | null;
+  projectKey?: ApplicationOidcForDevelopmentConfig['initialProjectKey'];
+  oAuthScopes?: ApplicationOidcForDevelopmentConfig['oAuthScopes'];
+  teamId?: ApplicationOidcForDevelopmentConfig['teamId'];
 };
 
-export const buildOidcScope = (options: BuilOidcScopeOptions): string => {
+const buildOidcScope = (options: BuilOidcScopeOptions): string => {
   const projectClaims = [];
   if (options.projectKey) {
     projectClaims.push(`${OIDC_CLAIMS.PROJECT_KEY}:${options.projectKey}`);
     projectClaims.push(
-      ...(options.permissions?.view ?? []).map(
+      ...(options.oAuthScopes?.view ?? []).map(
         (scope) => `${OIDC_CLAIMS.VIEW}:${scope}`
       )
     );
     projectClaims.push(
-      ...(options.permissions?.manage ?? []).map(
+      ...(options.oAuthScopes?.manage ?? []).map(
         (scope) => `${OIDC_CLAIMS.MANAGE}:${scope}`
       )
     );
@@ -38,3 +34,5 @@ export const buildOidcScope = (options: BuilOidcScopeOptions): string => {
     ...projectClaims,
   ].join(' ');
 };
+
+export { buildOidcScope };
