+++ packages/application-config/src/index.ts
@@ -1,2 +1,4 @@
 export { default as processConfig } from './process-config';
+export { buildOidcScope } from './build-oidc-scope';
 export * from './types';
+export * from './constants';
