+++ packages/application-shell/src/components/authenticated/has-cached-authentication-state.ts
@@ -1,8 +1,8 @@
 import type { ApplicationWindow } from '@commercetools-frontend/constants';
 
+import { buildOidcScope } from '@commercetools-frontend/application-config';
 import { STORAGE_KEYS } from '../../constants';
 import * as oidcStorage from '../../utils/oidc-storage';
-import { buildOidcScope } from './helpers';
 
 declare let window: ApplicationWindow;
 
@@ -34,7 +34,11 @@ const hasCachedAuthenticationState = (): boolean => {
         }
       }
       // Rebuild the requested OIDC scope to verify that it didn't change.
-      const requestedScope = buildOidcScope({ projectKey: activeProjectKey });
+      const requestedScope = buildOidcScope({
+        projectKey: activeProjectKey ?? undefined,
+        oAuthScopes: window.app.__DEVELOPMENT__?.oAuthScopes,
+        teamId: window.app.__DEVELOPMENT__?.teamId,
+      });
       // Check that the session scope didn't change.
       if (cachedScope === requestedScope) {
         return true;
