+++ playground/custom-application-config.json
@@ -3,7 +3,7 @@
   "entryPointUriPath": "state-machines",
   "cloudIdentifier": "${env:CLOUD_IDENTIFIER}",
   "mcApiUrl": "${env:MC_API_URL_GCP_STAGING}",
-  "permissions": {
+  "oAuthScopes": {
     "view": ["view_orders", "view_states"],
     "manage": ["manage_orders"]
   },
